feat(get-latest-workflow-artifact): consider runs triggered by comments
If a workflow is triggered by a comment, we have no clear way to
associate it with a PR through the head_sha. Instead we have to look at
the time and title of the run. This is not completely reliable but
should be "good enough" for most cases. Due to that it is disabled by
default.